Support ares-setup-device to APIs
:Release Notes:
Support ares-setup-device to APIs

:Detailed Notes:
- Refactoring code to support CLI and APIs
- showDeviceList, showDeviceListFill, resetDeviceList, setDefaultDevice,
  modifyDeviceInfo and removeDeviceInfoas APIs

:Testing Performed:
- Passed CLI unit test on OSE target
- Pass eslint
- Check each APIs are woking fine using test app

:Issues Addressed:
[WRP-2690] Develop APIs of ares-setup-device
