+++ lib/base/error-handler.js
@@ -81,6 +81,7 @@ const log = require('npmlog'),
             "INVALID_DEVICENAME" : "Invalid device name. The device name should consist of letters, numbers, and special characters ('-','_','#') and should start with letters or '_'",
             "INVALID_OBJECT" : "Object format error",
             "INVALID_MODE" : "Please specify an option, either '--add' or '--modify'",
+            "INVALID_REMOVE_MODE" : "Please set remove option",
             "INVALID_CAPTURE_FORMAT" : "Please specify the file extension(.png, .bmp or .jpg)",
             "INVALID_CSV_FORMAT" : "Please change the file extension to .csv",
             "INVALID_COMMAND" : "This command is invalid. Please check the supported commands using ares -l",
