+++ lib/base/setup-device.js
@@ -7,15 +7,34 @@
 const async = require('async'),
     chalk = require('chalk'),
     Table = require('easy-table'),
+    npmlog = require('npmlog'),
+    Appdata = require('./cli-appdata'),
+    errHndl = require('./error-handler'),
     novacom = require('./novacom');
 
 (function () {
-    const devicetools = {};
+    const log = npmlog;
+    log.heading = 'setup-device';
+    log.level = 'warn';
+
+    const appdata = new Appdata(),
+        devicetools = {},
+        defaultDeviceInfo = {
+            profile: appdata.getConfig(true).profile,
+            host: "127.0.0.1",
+            port: 22,
+            username: "root",
+            description: "new device description",
+            files: "stream",
+            default: false
+        };
+
     if (typeof module !== 'undefined' && module.exports) {
         module.exports = devicetools;
         module.exports.isValidDeviceName = isValidDeviceName;
         module.exports.isValidIpv4 = isValidIpv4;
         module.exports.isValidPort =isValidPort;
+        module.exports.replaceDefaultDeviceInfo = replaceDefaultDeviceInfo;
     }
 
     function isValidDeviceName(name) {
@@ -31,16 +50,175 @@ const async = require('async'),
         return String(port).match(/^[0-9]+$/);
     }
 
-    function finish(returnValue, next){
-        if(next && typeof next === 'function'){
-            if (!returnValue)
-                return setImmediate(next);
-            return setImmediate(next, null, returnValue);
-        }
-        return returnValue;
-    }
+    devicetools.showDeviceList = function(next){
+        async.waterfall([
+            _getDeviceInfo.bind(this, 'list'),
+            function(data, next){
+                const table = new Table();
+                data.forEach(function(item){
+                    if (!isValidDeviceName(item.name)) {
+                        return;
+                    }
+                    table.cell('name', (item.default === true) ? item.name + chalk.green(' (default)') : item.name );
+                    table.cell('deviceinfo', item.info);
+                    table.cell('connection', item.connection);
+                    table.cell('profile', item.profile);
+                    table.newRow();
+                });
+                next(null, table.toString());
+            }
+        ], function(err, results){
+            next(null,{msg: results});
+        });
+    };
 
-    function getDeviceInfo(mode, next){
+    devicetools.showDeviceListFull = function(next){
+        async.waterfall([
+            _getDeviceInfo.bind(this, 'full'),
+            function(data, next){
+                next(null, JSON.stringify(data, null, 4));
+            }
+        ], function(err, results){
+            next(null,{msg: results});
+        });
+    };
+
+    devicetools.resetDeviceList = function(next){
+        async.series([
+            function(next) {
+                appdata.resetDeviceList(next);
+            },
+            this.showDeviceList.bind(this, next)
+        ], function(err) {
+            next(err);
+        });
+    };
+
+    devicetools.setDefaultDevice = function(name, next){
+        try {
+            const resolver = this.resolver || (this.resolver = new novacom.Resolver()),
+                inDevice = {name: name, default: true};
+            async.series([
+                resolver.load.bind(resolver),
+                resolver.modifyDeviceFile.bind(resolver, 'default', inDevice),
+                this.showDeviceList.bind(this, next)
+            ], function(err) {
+                if (err) {
+                    return next(err);
+                }
+                next();
+            });
+        } catch (err) {
+            next(err);
+        }
+    };
+
+    devicetools.removeDeviceInfo = function(options, next){
+        try {
+            // For CLI
+            if (options.remove === 'true') {
+                next(errHndl.getErrMsg("EMPTY_VALUE", "DEVICE_NAME"));
+            }
+            // For API
+            if (!options.remove) {
+                next(errHndl.getErrMsg("INVALID_REMOVE_MODE"));
+            }
+
+            const resolver = this.resolver || (this.resolver = new novacom.Resolver()),
+                inDevice = {name: options.remove, profile: defaultDeviceInfo.profile};
+            async.series([
+                resolver.load.bind(resolver),
+                resolver.modifyDeviceFile.bind(resolver, 'remove', inDevice),
+                this.showDeviceList.bind(this, next)
+            ], function(err) {
+                if (err) {
+                    return next(err);
+                }
+                next();
+            });
+        } catch (err) {
+            next(err);
+        }
+    };
+
+    devicetools.modifyDeviceInfo = function(options, next){
+        try {
+            const mode = (options.add)? "add" : (options.modify)? "modify" : null;
+            if (!mode) {
+                return next(errHndl.getErrMsg("INVALID_MODE"));
+            }
+            if (options[mode].match(/^-/)) {
+                return next(errHndl.getErrMsg("EMPTY_VALUE", "DEVICE_NAME"));
+            }
+            const argName = (options.info)? "info" : mode;
+            const inDevice = _getParams(options, argName);
+            if (!inDevice.name) {
+                if (options[mode] === "true") {
+                    return next(errHndl.getErrMsg("EMPTY_VALUE", "DEVICE_NAME"));
+                }
+                inDevice.name = options[mode];
+            }
+
+            log.info("modifyDeviceInfo()", "devicename:", inDevice.name, ", mode:", mode);
+
+            if (inDevice.default !== undefined && mode === "modify") {
+                log.verbose("modifyDeviceInfo()", "Ignoring invalid arguments:default");
+                inDevice.default = undefined;
+            }
+
+            if (inDevice.privateKey) {
+                inDevice.privatekey = inDevice.privateKey;
+            }
+            if (typeof inDevice.privatekey === "string") {
+                inDevice.privateKey = inDevice.privatekey;
+                inDevice.privateKey = { "openSsh": inDevice.privateKey };
+                delete inDevice.privatekey;
+                inDevice.password = "@DELETE@";
+            }
+            if (typeof inDevice.password !== "undefined" && inDevice.password !== "@DELETE@") {
+                inDevice.privateKey = "@DELETE@";
+                inDevice.passphrase = "@DELETE@";
+            }
+
+            if (mode === "add") {
+                replaceDefaultDeviceInfo(inDevice);
+                if (!inDevice.privateKey && !inDevice.password) {
+                    inDevice.password = "";
+                }
+            }
+            // check validation
+            if (!isValidDeviceName(inDevice.name)) {
+                return next(errHndl.getErrMsg("INVALID_DEVICENAME"));
+            }
+            if (inDevice.host && !isValidIpv4(inDevice.host)) {
+                return next(errHndl.getErrMsg("INVALID_VALUE", "host", inDevice.host));
+            }
+            if (inDevice.port && !isValidPort(inDevice.port)) {
+                return next(errHndl.getErrMsg("INVALID_VALUE", "port", inDevice.port));
+            }
+            if (inDevice.port) {
+                inDevice.port = Number(inDevice.port);
+            }
+            if (!inDevice.profile) {
+                inDevice.profile = defaultDeviceInfo.profile;
+            }
+            const resolver = this.resolver || (this.resolver = new novacom.Resolver());
+            async.series([
+                resolver.load.bind(resolver),
+                resolver.modifyDeviceFile.bind(resolver, mode, inDevice),
+                this.showDeviceList.bind(this, next)
+            ], function(err) {
+                if (err) {
+                    return next(err);
+                }
+                next();
+            });
+        } catch (err) {
+            next(err);
+        }
+    };
+
+    function _getDeviceInfo(mode, next){
         const datas= [],
             resolver = new novacom.Resolver();
         async.waterfall([
@@ -86,45 +264,61 @@ const async = require('async'),
                         datas.push(data);
                     });
                 }
-                next(datas);
+                next(null, datas);
             }
-        ], function(err){
-                finish(err, next);
+        ], function(err, results){
+            next(err, results);
         });
     }
 
-    devicetools.showDeviceListFull = function(next){
-        async.waterfall([
-            getDeviceInfo.bind(this, 'full'),
-            function(data, next){
-                console.log(JSON.stringify(data, null, 4));
-                next();
-            }
-        ], function(err){
-            finish(err, next);
-        });
-    };
+    function _getParams(argv, option) {
+        let inputParams = [];
+        const params = {};
+        if (argv[option]) {
+            inputParams = [].concat(argv[option]);
+        }
 
-    devicetools.showDeviceList = function(next){
-        async.waterfall([
-            getDeviceInfo.bind(this, 'list'),
-            function(data, next){
-                const table = new Table();
-                data.forEach(function(item){
-                    if (!isValidDeviceName(item.name)) {
-                        return;
-                    }
-                    table.cell('name', (item.default === true) ? item.name + chalk.green(' (default)') : item.name );
-                    table.cell('deviceinfo', item.info);
-                    table.cell('connection', item.connection);
-                    table.cell('profile', item.profile);
-                    table.newRow();
-                });
-                console.log(table.toString());
-                next();
+        if (inputParams.length === 1 && inputParams[0].indexOf('{') !== -1 && inputParams[0].indexOf('}') !== -1 &&
+            ( (inputParams[0].split("'").length - 1) % 2) === 0) {
+            // eslint-disable-next-line no-useless-escape
+            inputParams[0] = inputParams[0].replace(/\'/g,'"');
+        }
+        inputParams.forEach(function(strParam) {
+            try {
+                const data = JSON.parse(strParam);
+                for (const k in data) {
+                    params[k] = data[k];
+                }
+            } catch (err) {
+                const tokens = strParam.split('=');
+                if (tokens.length === 2) {
+                    params[tokens[0]] = tokens[1];
+                    log.verbose("_getParams()", "Inserting params ", tokens[0] + " = " + tokens[1]);
+                } else {
+                    log.verbose("_getParams()", "Ignoring invalid arguments:", strParam);
+                }
             }
-        ], function(err){
-            finish(err, next);
         });
-    };
+
+        // FIXME : -i default=true is set as "true" string
+        if (params.default !== undefined && typeof params.default == "string") {
+            params.default = (params.default === "true");
+        }
+
+        log.silly("_getParams()", "params:", JSON.stringify(params));
+        return params;
+    }
+
+    function replaceDefaultDeviceInfo(inDevice) {
+        if (inDevice) {
+            inDevice.profile = inDevice.profile || defaultDeviceInfo.profile;
+            inDevice.type = inDevice.type || defaultDeviceInfo.type;
+            inDevice.host = inDevice.host || defaultDeviceInfo.host;
+            inDevice.port = inDevice.port || defaultDeviceInfo.port;
+            inDevice.username = inDevice.username || defaultDeviceInfo.username;
+            inDevice.files = inDevice.files || defaultDeviceInfo.files;
+            inDevice.description = inDevice.description || defaultDeviceInfo.description;
+            inDevice.default = inDevice.default || defaultDeviceInfo.default;
+        }
+    }
 }());
